backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.129.1
filelock==3.24.3
fsspec==2026.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
openai-whisper==20250625
pillow==12.0.0
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
regex==2026.2.19
requests==2.32.5
sounddevice==0.5.5
starlette==0.52.1
sympy==1.14.0
tiktoken==0.12.0
torch==2.10.0
torchaudio==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==16.0
